+++ app/src/main/java/com/example/dmsport_android/network/MyPageApi.kt
@@ -1,0 +1,13 @@
+package com.example.dmsport_android.network
+
+import com.example.dmsport_android.dto.response.MyPageResponse
+import retrofit2.Response
+import retrofit2.http.GET
+import retrofit2.http.Header
+
+interface MyPageApi {
+    @GET("users/my")
+    suspend fun my(
+        @Header("Authorization") accessToken : String,
+    ) : Response<MyPageResponse>
+}
