+++ app/src/main/java/com/example/dmsport_android/repository/RegisterRepository.kt
@@ -11,19 +11,19 @@ import retrofit2.Response
 class RegisterRepository {
 
     suspend fun duplicate(duplicateRequest: DuplicateRequest) : Response<Void> {
-        return ApiProvider.retrofit.duplicate(duplicateRequest)
+        return ApiProvider.registerApi.duplicate(duplicateRequest)
     }
 
     suspend fun verifyEmail(verifyEmailRequest : VerifyEmailRequest) : Response<Void>{
-        return ApiProvider.retrofit.verifyEmail(verifyEmailRequest)
+        return ApiProvider.registerApi.verifyEmail(verifyEmailRequest)
     }
 
     suspend fun verify(verifyRequest: VerifyRequest) : Response<Void>{
-        return ApiProvider.retrofit.verify(verifyRequest)
+        return ApiProvider.registerApi.verify(verifyRequest)
     }
 
     suspend fun register(registerRequest: RegisterRequest) : Response<RegisterResponse> {
-        return ApiProvider.retrofit.register(registerRequest)
+        return ApiProvider.registerApi.register(registerRequest)
     }
 
 }
